requirements.txt
numpy==1.26.4
pandas==2.2.1
python-Levenshtein==0.25.0
rouge-score==0.1.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1